Convert EX to register the inputs rather than outputs
This was the last step to convert.  The motivation for this is
primarily to make it easier to improve timing as controlling fan-out
is one of the key things to worry about and having registers at the
output of a module makes this harder to see/reason about.

Registered inputs to each stage also better matches the way I tend to
think about it.