[supervisor] Added most of the supervisor CSRs and improved the exception handling
RISC-V is actually rather involved in this area and this is still not
complete, thus the changes don't show in regression, but in reality
rv32mi-p-illegal got a bit farther.

Premature optimization yadda yadda, but the exception/CSR handling
sure looks like it might have a timing impact unless pipelined
somehow.