[restructure] subsume DE/RF into EX
It turned out pretty awkward to have a module per stage because
there's a lot of wiring between them and it's makes it a lot of work
to move logic around.  Instead, the path going forward is to try to
factor out self-contained combinatorial blocks into their own modules,
keeping the backend stages inside one module.

As a consequence, ME is next.