YAGS + new license: 98.1 DMIPS, 0.904 (+7%) IPC, 56.6 MHz
Added a YAGS history based correction to the Bimodal predictor which
means *NO* control-flow mispredictions in the Dhrystone loop (after a
few iterations as the BP is trained).  This improves IPC and the
remaining bubbles are due to load-use hazards and one load-hit-store
case.  Unfortunate this hurts timing which offsets the IPC gain.

The other big change is the move to the ISC license (same as
PicoRV32).