Reduce redundant build code, pass initial memory to simulator
The key desire here is to be able to reuse the built simulator on
multiple workloads.  It's not fully complete as compliance tests has
further dependencies on the signature and I haven't dealt with that.

We use $value$argsplus for this, but that isn't supported by Yosys it
seems.

Also, while at it, making executables in .elf makes build rules
simpler.